# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Card Management System API
  version: "1.0"
  description: API for managing bank cards, users, and transfers
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateUserRequest:
      type: object
      required: [name, surname, birthDate, email, password, role]
      properties:
        name:
          type: string
        surname:
          type: string
        birthDate:
          type: string
          format: date
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [ADMIN, USER]
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        surname:
          type: string
        birthDate:
          type: string
          format: date
        email:
          type: string
        role:
          type: string
...
paths:
...
  /api/admin/users:
    post:
      tags: [Admin Users]
      summary: Create a new user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    get:
      tags: [Admin Users]
      summary: Get all users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
  /api/admin/users/{id}:
    get:
      tags: [Admin Users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    delete:
      tags: [Admin Users]
      summary: Delete user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
  /api/admin/users/by-email:
    get:
      tags: [Admin Users]
      summary: Get user by email
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
...